+++ docs/release-notes.md
@@ -1,5 +1,12 @@
 # Release notes
 
+## 1.0.3
+
+- Added the ability to transform between datasets that are indirectly registered,
+  for example two zoom datasets that are registered to the same common overview dataset.
+- Added [hoa_tools.dataset.Dataset.get_registered][] to get all datasets registered
+  (directly or indirectly) to a given dataset.
+
 ## 1.0.2
 
 - Fixed the inclusion of the metadata inventory in the distribution on PyPI.